requirements.txt
gradio
torch
torchvision
segmentation-models-pytorch
albumentations
opencv-python-headless
numpy
Pillow
